Move autoalloc fragments to its screen module
