--- crates/hyperqueue/src/dashboard/ui/fragments/auto_allocator/mod.rs
@@ -1,5 +1,0 @@
-pub mod alloc_timeline_chart;
-pub mod allocations_info_table;
-pub mod fragment;
-pub mod queue_info_table;
-pub mod queue_params_display;
