--- crates/hyperqueue/src/dashboard/ui/fragments/mod.rs
@@ -1,1 +1,0 @@
-pub mod auto_allocator;
