+++ crates/hyperqueue/src/dashboard/ui/mod.rs
@@ -1,4 +1,3 @@
-pub mod fragments;
 pub mod screen;
 pub mod screens;
 pub mod styles;
