+++ crates/hyperqueue/src/dashboard/ui/screens/autoalloc/alloc_timeline_chart.rs
