+++ crates/hyperqueue/src/dashboard/ui/screens/autoalloc/allocations_info_table.rs
