+++ crates/hyperqueue/src/dashboard/ui/screens/autoalloc/fragment.rs
@@ -9,10 +9,10 @@ use std::time::SystemTime;
 
 use crate::dashboard::data::timelines::alloc_timeline::AllocationQueueInfo;
 use crate::dashboard::data::DashboardData;
-use crate::dashboard::ui::fragments::auto_allocator::alloc_timeline_chart::AllocationsChart;
-use crate::dashboard::ui::fragments::auto_allocator::allocations_info_table::AllocationInfoTable;
-use crate::dashboard::ui::fragments::auto_allocator::queue_info_table::AllocationQueueInfoTable;
-use crate::dashboard::ui::fragments::auto_allocator::queue_params_display::QueueParamsTable;
+use crate::dashboard::ui::screens::autoalloc::alloc_timeline_chart::AllocationsChart;
+use crate::dashboard::ui::screens::autoalloc::allocations_info_table::AllocationInfoTable;
+use crate::dashboard::ui::screens::autoalloc::queue_info_table::AllocationQueueInfoTable;
+use crate::dashboard::ui::screens::autoalloc::queue_params_display::QueueParamsTable;
 use crate::server::autoalloc::QueueId;
 use ratatui::layout::{Constraint, Direction, Layout, Rect};
 
@@ -104,7 +104,7 @@ impl AutoAllocatorFragment {
             FocusedComponent::AllocationInfoTable => self.allocations_info_table.handle_key(key),
         };
 
-        match key.code{
+        match key.code {
             KeyCode::Char('1') => {
                 self.component_in_focus = FocusedComponent::QueueParamsTable;
                 self.allocations_info_table.clear_selection();
