+++ crates/hyperqueue/src/dashboard/ui/screens/autoalloc/queue_info_table.rs
